Comments applied. Delete Folder updated.

+++ RedNimbus/BucketService/Services/FileSystemService.cs
@@ -1,4 +1,5 @@
-﻿using System;
+﻿using RedNimbus.BucketService.Helper;
+using System;
 using System.Collections.Generic;
 using System.IO;
 
@@ -6,7 +7,44 @@ namespace RedNimbus.BucketService.Services
 {
     public static class FileSystemService
     {
+        /*
+        public static bool CreateUser(string path)
+        {
+            try
+            {
+                string pathForZfs = path.TrimStart('/');
+                string cmd = "zfs create " + pathForZfs;
+                ShellHelper.Bash(cmd);
+                return true;
+            }
+            catch(Exception)
+            {
+                return false;
+            }
+        }
 
+        public static bool CreateBucket(string path)
+        {
+            try
+            {
+                string pathForZfs = path.TrimStart('/');
+                string cmd = "zfs create " + pathForZfs;
+                ShellHelper.Bash(cmd);
+                cmd = "zfs set quota=5G " + pathForZfs;
+                ShellHelper.Bash(cmd);
+                return true;
+            }
+            catch (Exception)
+            {
+                return false;
+            }
+        }
+        */
+
+        public static int NumberOfDirectories(string path)
+        {
+            return Directory.GetDirectories(path).Length;
+        }
 
         public static List<string> ListContent(string path)
         {
@@ -15,7 +53,7 @@ namespace RedNimbus.BucketService.Services
             {
                 foreach (string entry in Directory.GetDirectories(path))
                 {
-                    //TODO: Change this shit
+                    //TODO: Find better solution for this, later.
                     string[] val = entry.Split('/');
                     string last = val[val.Length - 1];
                     string[] splitLast = last.Split('\\');
@@ -24,7 +62,7 @@ namespace RedNimbus.BucketService.Services
                 returnValue.Add("*");
                 foreach (string entry in Directory.GetFiles(path))
                 {
-                    //TODO: Change this shit
+                    //TODO: Find better solution for this, later.
                     string[] val = entry.Split('/');
                     string last = val[val.Length - 1];
                     string[] splitLast = last.Split('\\');
@@ -35,11 +73,10 @@ namespace RedNimbus.BucketService.Services
             {
                 return null;
             }
-            finally { }
             return returnValue;
         }
 
-        public static bool CreateBucket(string path)
+        public static bool CreateFolder(string path)
         {
             try
             {
@@ -56,24 +93,51 @@ namespace RedNimbus.BucketService.Services
             {
                 return false;
             }
-            finally { }
 
         }
 
-        public static bool DeleteBucket(string path)
+        public static bool DeleteFolder(string path)
         {
             try
             {
+                bool successful = true;
+                foreach (string entry in Directory.GetFiles(path))
+                {
+                    successful &= DeleteFile(entry);
+                }
+                foreach (string entry in Directory.GetDirectories(path))
+                {
+                    successful &= DeleteFolder(entry);
+                }
                 // Try to delete the directory.
-                Directory.Delete(path);
-                return true;
+                if (CheckFolderEmpty(path))
+                {
+                    Directory.Delete(path);
+                    return successful;
+                }
+                return false;
             }
             catch (Exception)
             {
                 return false;
             }
-            finally { }
 
+        }
+
+        public static bool CheckFolderEmpty(string path)
+        {
+            if (string.IsNullOrEmpty(path))
+            {
+                throw new ArgumentNullException("path");
+            }
+
+            var folder = new DirectoryInfo(path);
+            if (folder.Exists)
+            {
+                return folder.GetFileSystemInfos().Length == 0;
+            }
+
+            throw new DirectoryNotFoundException();
         }
 
         public static bool ByteArrayToFile(string path, byte[] fileAsByteArray)
@@ -87,7 +151,6 @@ namespace RedNimbus.BucketService.Services
             {
                 return false;
             }
-            finally { }
         }
 
         public static byte[] FileToByteArray(string path)
@@ -100,7 +163,6 @@ namespace RedNimbus.BucketService.Services
             {
                 return null;
             }
-            finally { }
         }
 
         public static bool DeleteFile(string path)
@@ -114,7 +176,6 @@ namespace RedNimbus.BucketService.Services
             {
                 return false;
             }
-            finally { }
         }
     }
 }
